AiAccountingAgent/pyproject.toml
[project]
name = "aiaccountingagent"
version = "0.1.0"
description = "AI agent for Tripletex accounting tasks — NM i AI 2026"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "requests>=2.32",
    "pydantic>=2.7",
    "google-genai>=1.0",
    "pypdf>=4.0",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
